料理集約ルートに別の集約ルートのレシピ元をメソッドで追加する（今できるところ全部GraphQL反映） (#46)
* graphqlのテストで、テスト失敗時にレスポンスがrailsの開発者モードのエラーページHTMLだった場合、tmpディレクトリに保存するよう修正

* Mutations::Dish::AddDishに新Command適用

* Business::Food::Dish::Usecase::AddWithNewSourceCommandの返り値をdishだけでなくsourceも

* Mutations::Dish::AddDishWithNewSourceに新Command適用

* Mutations::Dish::UpdateDishに新Command適用

* Business::Food::Dish::Usecase::UpdateWithNewSourceCommandの返り値をdishだけでなくsourceも

* Mutations::Dish::UpdateDishWithNewSourceに新Command適用

* Mutations::Dish::RemoveDishに新Command適用

+++ backend/app/domain/business/food/dish/usecase/remove_command.rb
@@ -10,7 +10,8 @@ module Business::Food::Dish
       dish_record = ::Dish.find_by(id: dish_id, user_id: user_id)
       raise "指定した料理は存在しません。" if dish_record.blank?
 
-      # TODO: 紐付く食事がある場合削除できないチェック作る
+      # ユースケースがはっきり定まっていないので、定まるまで安全に倒す
+      raise "この料理は登録されている食事があるので削除できません。" if dish_record.meals.present?
 
       dish_record.destroy!
     end
